added 100% solution to exam prep task Meme Lounge

+++ previous-exams/5. JS Applications Exam Preparation - Meme Lounge/solution/src/views/delete.js
@@ -1,8 +1,13 @@
 import { deleteMeme } from "../api/data.js";
 
 
-export function viewDelete(ctx) {
-    deleteMeme(ctx.params.id);
+export async function viewDelete(ctx) {
 
-    ctx.page.redirect('/');
+    const choice = confirm('Are you sure you want to delete this entry?')
+
+    if (choice) {
+        await deleteMeme(ctx.params.id);
+
+        ctx.page.redirect('/');
+    }
 }
